Zero Touch Provisioning Support Stage 1 (#1070)
Starts us down the path of fixing https://github.com/the-draupnir-project/Draupnir/issues/1023

This PR gets us Zero Touch Provisioning support for AS mode and Bot mode. If your env can take usr and pswd and turn it into access tokens or however you want to do that then well we dont need to do PSWD Auth at all.

If you dont PSWD auth is useful for this and is probably what i will resort to in mdad for Zero Touch Provisioning.

Draft as this PR is very much not even mx-tested because mx-tester decided to say nope cat.

* Zero Touch Provisioning Support

* Update Account data keys to Gnuxie Suggested Values


* Refine ZTD


* Refine Config validation

* Run Prettier

* Fix Test Linting

* Refine ZTD Branch by removing leftover config values.

* Remove fallback management room value as it breaks ZTP

* Fix config validation having Truthy problems

* Fix dangling import.

* Fix config validation error.

* Revert "Fix config validation error."

This reverts commit c313dcbb529da7e79387445b20b49a84f9f8d037.

* Remove fake cast on config.initialManager

* note only available in develop while we fix shit

* Stop tests from accessing config.managementRoom directly.

---------

+++ apps/draupnir/test/appservice/utils/harness.ts
@@ -1,3 +1,4 @@
+// SPDX-FileCopyrightText: 2026 Catalan Lover <catalanlover@protonmail.com>
 // Copyright 2022 Gnuxie <Gnuxie@protonmail.com>
 // Copyright 2022 The Matrix.org Foundation C.I.C.
 // SPDX-FileCopyrightText: 2026 Catalan Lover <catalanlover@protonmail.com>
@@ -33,33 +34,38 @@ export async function setupHarness(): Promise<MjolnirAppService> {
 }
 
 export async function setupHarnessWithConfig(
-  config: AppserviceConfig
+  config: AppserviceConfig,
+  options: {
+    ensureAdminRoomAlias?: boolean;
+  } = {}
 ): Promise<MjolnirAppService> {
-  const utilityUser = await newTestUser(config.homeserver.url, {
-    name: { contains: "utility" },
-    isAdmin: true,
-  });
-  if (
-    typeof config.adminRoom !== "string" ||
-    !isStringRoomAlias(config.adminRoom)
-  ) {
-    throw new TypeError(
-      "This test expects the harness config to have a room alias."
+  if (options.ensureAdminRoomAlias ?? true) {
+    const utilityUser = await newTestUser(config.homeserver.url, {
+      name: { contains: "utility" },
+      isAdmin: true,
+    });
+    if (
+      typeof config.adminRoom !== "string" ||
+      !isStringRoomAlias(config.adminRoom)
+    ) {
+      throw new TypeError(
+        "This test expects the harness config to have a room alias."
+      );
+    }
+    await utilityUser
+      .deleteRoomAlias(config.adminRoom)
+      // Ignore if alias doesn't exist yet.
+      .catch((_: unknown) => undefined);
+    const adminRoomID = await ensureAliasedRoomExists(
+      utilityUser,
+      config.adminRoom
+    );
+    await utilityUser.setUserPowerLevel(
+      "@draupnir-moderation:localhost:9999",
+      adminRoomID,
+      100
     );
   }
-  await utilityUser
-    .deleteRoomAlias(config.adminRoom)
-    // Ignore if alias doesn't exist yet.
-    .catch((_: unknown) => undefined);
-  const adminRoomID = await ensureAliasedRoomExists(
-    utilityUser,
-    config.adminRoom
-  );
-  await utilityUser.setUserPowerLevel(
-    "@draupnir-moderation:localhost:9999",
-    adminRoomID,
-    100
-  );
   return await MjolnirAppService.run(
     9000,
     config,
